refactor(core): 迁移 ModuleHelper 至 app.core.module_loader,消除 core->helper.module 反向依赖
ModuleHelper 是仅依赖标准库与 app.log 的动态模块加载工具,原置于 helper 层,造成
core.module -> helper.module 的反向依赖。现迁移到正确的 core 层(内容不变),
app/helper/module.py 改为 re-export 垫片,使既有导入路径(含插件 autosignin)零改动可用。

- git mv app/helper/module.py -> app/core/module_loader.py(内容不变)
- 更新 4 个内部导入方:core/module、workflow、modules/indexer、modules/filemanager
- 插件 app.plugins.autosignin 经垫片继续工作,无需改动(已验证解析到同一类)
- 新增 tests/test_core_module_relocation.py:垫片同类(含插件路径)/ load 功能冒烟 /
  core.module 不再引用 helper.module / module_loader 无 helper 依赖
