+++ app/modules/indexer/__init__.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Tuple, Union
 
 from app.core.context import SubtitleInfo, TorrentInfo
 from app.db.site_oper import SiteOper
-from app.helper.module import ModuleHelper
+from app.core.module_loader import ModuleHelper
 from app.helper.sites import SitesHelper  # noqa
 from app.log import logger
 from app.modules import _ModuleBase
